# Use a faster and safer implementation of `alias_sample!`