@devmotion 2024-04-12
AFAIK there's no standard rule to follow. I just wanted to wait a few days in case someone has additional comments, but it seems nobody objected to the PR. IMO the PR is a clear improvement so I'll merge and leave further improvements or comments for future PRs.